refactor(core): move HStack, VStack, Heading, Code into own source dirs
Fixes #2644
