+++ packages/core/src/Code/XDSCode.tsx
@@ -6,7 +6,7 @@
  * @file XDSCode.tsx
  * @input Uses React, StyleX, theme tokens
  * @output Exports XDSCode component for inline code styling
- * @position Core implementation; inline code display
+ * @position Core implementation; lives in own Code/ dir, re-exported by CodeBlock/
  *
  * SYNC: When modified, update:
  * - /packages/core/src/CodeBlock/index.ts (exports if types change)
